+++ packages/ntcore-ts-client/src/index.ts
@@ -1,1 +1,3 @@
 export * from './lib/ntcore-ts-client';
+export * from './lib/pubsub/topic';
+export * from './lib/types/types';