+++ packages/ntcore-ts-client/src/lib/ntcore-ts-client.ts
@@ -1,8 +1,8 @@
 import { PubSubClient } from './pubsub/pubsub';
-import { Topic } from './pubsub/topic';
+import { NetworkTablesTopic } from './pubsub/topic';
 import { Util } from './util/util';
 
-import type { NetworkTableTypeInfo, NetworkTableTypes } from './types/types';
+import type { NetworkTablesTypeInfo, NetworkTableTypes } from './types/types';
 
 /** Properties for creating the NetworkTables class. */
 interface NT_PROPS {
@@ -167,7 +167,7 @@ export class NetworkTables {
    * @param defaultValue - The default value of the topic.
    * @returns The topic.
    */
-  createTopic<T extends NetworkTableTypes>(name: string, typeInfo: NetworkTableTypeInfo, defaultValue?: T) {
-    return new Topic<T>(this._client, name, typeInfo, defaultValue);
+  createTopic<T extends NetworkTableTypes>(name: string, typeInfo: NetworkTablesTypeInfo, defaultValue?: T) {
+    return new NetworkTablesTopic<T>(this._client, name, typeInfo, defaultValue);
   }
 }