+++ packages/ntcore-ts-client/src/lib/util/util.ts
@@ -2,7 +2,7 @@ import { v4 as uuidv4 } from 'uuid';
 
 import { NetworkTableTypeInfos } from '../types/types';
 
-import type { BinaryMessage, NetworkTableTypeInfo, NetworkTableTypes, TypeNum, TypeString } from '../types/types';
+import type { BinaryMessage, NetworkTablesTypeInfo, NetworkTableTypes, TypeNum, TypeString } from '../types/types';
 
 /**
  * Class for holding utility functions.
@@ -39,7 +39,7 @@ export class Util {
    * @param data - The value.
    * @returns The NT type number.
    */
-  static getNetworkTableTypeFromObject(data: NetworkTableTypes): NetworkTableTypeInfo {
+  static getNetworkTableTypeFromObject(data: NetworkTableTypes): NetworkTablesTypeInfo {
     if (typeof data === 'boolean') {
       return NetworkTableTypeInfos.kBoolean;
     } else if (typeof data === 'number') {
@@ -137,7 +137,7 @@ export class Util {
     topicId: number,
     timestamp: number,
     data: NetworkTableTypes,
-    typeInfo?: NetworkTableTypeInfo
+    typeInfo?: NetworkTablesTypeInfo
   ): BinaryMessage {
     const type = typeInfo ?? this.getNetworkTableTypeFromObject(data);
     return [topicId, timestamp, type[0], data];