+++ packages/ntcore-ts-client/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.0.1-beta.4
+
+- Note: beta.1-beta.3 were tests of publishing to NPM and small fixes
+- Fixed dependency issue
+- Added keywords
+- Breaking changes:
+  - renamed Topic to NetworkTablesTopic
+  - renamed NetworkTableTypeInfo to NetworkTablesTypeInfo
+  - Moved to ES5/CommonJS compiling
+
 ## 0.0.1-beta.0
 
 - Initial Release